Task-3 - Rename editModeActionCreators to areEditableActionCreators
